# config-control-service/spec/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Config Control Service API",
    "description": "Centralized configuration management and drift detection service.\n\n**Features:**\n- Service instance heartbeat tracking\n- Configuration drift detection and reporting\n- Service discovery integration with Consul\n- Config refresh orchestration via Kafka\n- Real-time monitoring and alerts\n\n**Architecture:**\n- Hexagonal architecture with distinct API, Application, Domain, and Infrastructure layers\n- MongoDB for persistence\n- Redis for caching\n- Kafka for event broadcasting\n",
    "contact": {
      "name": "Platform Team",
      "email": "platform@example.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "development server"
    }
  ],
  "security": [
    {
      "oauth2_auth_code": [],
      "oauth2_password": []
    }
  ],
...
  "paths": {
...
    "/api/approval-requests": {
      "get": {
        "tags": [
          "Approval Requests"
        ],
        "summary": "List approval requests",
        "description": "Retrieve a paginated list of approval requests.\n\n**Access Control:**\n- Team members: Can view their own requests\n- SYS_ADMIN: Can view all requests\n- LINE_MANAGER: Can view requests from their direct reports\n- Results are automatically filtered based on user permissions\n",
        "operationId": "findAllApprovalRequests",
        "parameters": [
          {
            "name": "requesterUserId",
            "in": "query",
            "required": false,
            "schema": {
              "maxLength": 100,
              "minLength": 0,
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "maxLength": 20,
              "minLength": 0,
              "type": "string"
            }
          },
          {
            "name": "requestType",
            "in": "query",
            "required": false,
            "schema": {
              "maxLength": 50,
              "minLength": 0,
              "type": "string"
            }
          },
          {
            "name": "fromDate",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "toDate",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Zero-based page index (0..N)",
            "required": false,
            "schema": {
              "minimum": 0,
              "type": "integer",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "The size of the page to be returned",
            "required": false,
            "schema": {
              "minimum": 1,
              "type": "integer",
              "default": 20
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized - Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "Approval requests retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalRequestPageResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "oauth2_auth_code": []
          },
          {
            "oauth2_password": []
          }
        ]
      },
      "post": {
        "tags": [
          "Approval Requests"
        ],
        "summary": "Create approval request",
        "description": "Create a new approval request for service ownership.\n\n**Multi-Gate Approval Flow:**\n- SYS_ADMIN: Can approve/reject any request\n- LINE_MANAGER: Can approve requests from their direct reports\n- Request requires approval from configured gates before service ownership is transferred\n\n**Access Control:**\n- Team members: Can create requests for services not owned by their team\n- SYS_ADMIN: Can create requests for any service\n",
        "operationId": "createApprovalRequest",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalRequestCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "401": {
            "description": "Unauthorized - Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "201": {
            "description": "Approval request created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalRequestResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict - Request already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "oauth2_auth_code": []
          },
          {
            "oauth2_password": []
          }
        ]
      }
    },
    "/api/approval-requests/{id}/decisions": {
      "post": {
        "tags": [
          "Approval Requests"
        ],
        "summary": "Submit approval decision",
        "description": "Submit a decision (approve/reject) for an approval request.\n\n**Multi-Gate Approval:**\n- SYS_ADMIN: Can approve/reject any request\n- LINE_MANAGER: Can approve/reject requests from their direct reports\n- Request status is updated based on required approvals from configured gates\n\n**Access Control:**\n- Only authorized approvers can submit decisions\n- Decision is recorded with approver information and timestamp\n",
        "operationId": "submitApprovalDecision",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Approval request ID",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "request-12345"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalRequestDecisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "403": {
            "description": "Forbidden - Not authorized to approve this request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "Decision submitted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalRequestResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid decision data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Approval request not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "oauth2_auth_code": []
          },
          {
            "oauth2_password": []
          }
        ]
      }
    },
...
    "/api/approval-decisions/request/{requestId}": {
      "get": {
        "tags": [
          "Approval Decisions"
        ],
        "summary": "List decisions for a request",
        "description": "Retrieve all approval decisions for a specific approval request.\n\n**Access Control:**\n- Users can view decisions for requests they have access to\n- SYS_ADMIN can view decisions for any request\n",
        "operationId": "findApprovalDecisionsByRequestId",
        "parameters": [
          {
            "name": "requestId",
            "in": "path",
            "description": "Approval request ID",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "request-12345"
          },
          {
            "name": "page",
            "in": "query",
            "description": "Zero-based page index (0..N)",
            "required": false,
            "schema": {
              "minimum": 0,
              "type": "integer",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "The size of the page to be returned",
            "required": false,
            "schema": {
              "minimum": 1,
              "type": "integer",
              "default": 20
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized - Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "Approval decisions retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalDecisionPageResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "oauth2_auth_code": []
          },
          {
            "oauth2_password": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApprovalRequestPageResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "description": "List of approval requests in current page",
            "items": {
              "$ref": "#/components/schemas/ApprovalRequestResponse"
            }
          },
          "metadata": {
            "$ref": "#/components/schemas/PageMetadata"
          }
        },
        "description": "Page response for approval requests"
      },
      "ApprovalDecisionPageResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "description": "List of approval decisions in current page",
            "items": {
              "$ref": "#/components/schemas/ApprovalDecisionResponse"
            }
          },
          "metadata": {
            "$ref": "#/components/schemas/PageMetadata"
          }
        },
        "description": "Page response for approval decisions"
      },
...
    },
    "securitySchemes": {
      "oauth2_auth_code": {
        "type": "oauth2",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "http://keycloak:8080/realms/config-control/protocol/openid-connect/auth",
            "tokenUrl": "http://keycloak:8080/realms/config-control/protocol/openid-connect/token",
            "scopes": {
              "openid": "OpenID",
              "profile": "Profile",
              "email": "Email"
            }
          }
        }
      },
      "oauth2_password": {
        "type": "oauth2",
        "flows": {
          "password": {
            "tokenUrl": "http://keycloak:8080/realms/config-control/protocol/openid-connect/token",
            "scopes": {
              "openid": "OpenID",
              "profile": "Profile",
              "email": "Email"
            }
          }
        }
      }
    }
  }
}